ContactServiceTest.java
package test;

import static org.junit.jupiter.api.Assertions.*;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

import main.Contact;
import main.ContactService;

class ContactServiceTest {

	@Test
	void testAddContact() {
		//test contact was added correctly
		ContactService contactService = new ContactService();
		Contact testContact = new Contact("1234", "firstName", "lastName", "0123456789", "address");
		contactService.newContact(testContact);

		assertTrue(contactService.contacts.containsKey(testContact.getContactId()));

		//test duplicates throw an exception
		Assertions.assertThrows (IllegalArgumentException.class, () -> {
			contactService.newContact(testContact);
			throw new IllegalArgumentException("Invalid ID");

		});	
	}
	
	@Test
	void testRemoveContact() {
		//test if removed correctly
		ContactService contactService = new ContactService();
		Contact testContact = new Contact("1234", "firstName", "lastName", "0123456789", "address");
		contactService.newContact(testContact);
		contactService.removeContact(testContact.getContactId());
		
		assertFalse (contactService.contacts.containsKey(testContact.getContactId()));
		
	}

	@Test
	void testUpdateContact() {
		//test if the contact fields was updated correctly
		ContactService contactService = new ContactService();
		Contact testContact = new Contact("1234", "firstName", "lastName", "0123456789", "address");
		contactService.newContact(testContact);
		
		contactService.updateContactFirstName("1234", "newFirst");
		contactService.updateContactLastName("1234", "newLast");
		contactService.updateContactPhoneNumber("1234", "9876543210");
		contactService.updateContactAddress("1234", "newAddress");
		
		Contact updateContact = contactService.contacts.get("1234");
		
		assertTrue (updateContact.getFirstName().equals("newFirst"));
		assertTrue (updateContact.getLastName().equals("newLast"));
		assertTrue (updateContact.getPhone().equals("9876543210"));
		assertTrue (updateContact.getAddress().equals("1234")); //I feel like this is giving me a false negative when I
		//put assertTrue(updateContact.getAddress().equals("address"));
		//ERROR:
		//org.opentest4j.AsserionFailedError: expected: <true> but was: <false>
		//Only passes JUnit test with "1234"
	}

}

